requirements.txt
discord.py==2.4.0
google-cloud-firestore==2.16.0
google-generativeai==0.7.2
python-dotenv==1.0.1
